refactor(tui): move MultiPaneTui into its own module
The widget and its whole inherent impl leave the module root. The two
agent timing constants the rest of the crate reads stay on the root and
gain a line of documentation.
